requirements.txt
openai
langfuse>=2.0.0,<3.0.0
psycopg2-binary
pgvector
python-dotenv
rich
rank_bm25
